Handle null items of the report table better
Previously, if a SQL query did not return results for all rows, the results
that were returned would be added to the table in a new column(s) and those
(new) columns would be populated from top down. Therefore, the data would
not correcly line up with the row to which it pertains. Now, results are
matched back to the row to which they pertain, and rows are checked after
each query to ensure they all have the same number of columns.